moved part of the code from index.js to makeRequests func
